% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/couleurs.r
\name{couleurs_ipp}
\alias{couleurs_ipp}
\title{Renvoie le nom et le code hexadécimal des couleurs ipp.}
\usage{
couleurs_ipp(...)
}
\arguments{
\item{...}{une str ou un intérable de str}
}
\value{
Une liste str:str list de noms de couleurs et de codes hexadécimaux
}
\description{
Renvoie le nom et le code hexadécimal des couleurs ipp.
}
\examples{
couleurs_ipp['bleu_fonce']
}
