% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.r
\name{palette_ipp}
\alias{palette_ipp}
\title{Renvoie une palette}
\usage{
palette_ipp(palette = "principale", reverse = FALSE, ...)
}
\arguments{
\item{palette}{Nom de la palette définie dans definition_palettes_ipp}

\item{reverse}{Booléen pour inverser l'ordre des couleurs}

\item{...}{Paramètres passés à grDevices::colorRampPalette}
}
\value{
Une fonction qui à un entier associe une palette avec ce nombre de couleurs
}
\description{
Renvoie une palette
}
