% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales.r
\name{scale_fill_ipp}
\alias{scale_fill_ipp}
\title{Renvoie une échelle de remplissage pour ggplot2}
\usage{
scale_fill_ipp(palette = "principale", discrete = TRUE, reverse = FALSE, ...)
}
\arguments{
\item{palette}{Nom de la palette définie dans definition_palettes_ipp}

\item{reverse}{Booléen pour rendre l'échelle discrète ou continue}

\item{...}{Paramètres passés à ggplot2::scale_color_gradientn ou ggplot2::discrete_scale}
}
\value{
Une échelle de remplissage pour ggplot2
}
\description{
Renvoie une échelle de remplissage pour ggplot2
}
